+++ changelog.md
@@ -1,3 +1,7 @@
+# 0.10.3
+
+* Redefining lifecycle hooks will now automatically compose them, implements [#252](https://github.com/mobxjs/mobx-state-tree/issues/252)
+
 # 0.10.2
 
 * Object model instances no longer share a prototype.